+++ NEWS.md
@@ -1,3 +1,16 @@
+### Version 1.2 (2018-10-11)
+
+- Changes to the Lattes website seems to be permanent and stable. The main change is that in order to download  xml zip files, one must mannualy break a captcha. Based on this change, the package no longer works by downloading the files, but acessing it from locally (yes, you must download all files manualy). This update fix [Github issue 10](https://github.com/msperlin/GetLattesData/issues/10) and [Github issue 09](https://github.com/msperlin/GetLattesData/issues/09).
+
+
+### Version 1.0 (2018-08-13)
+
+Lattes website is back without captcha! The new version of GetLattesData is a bit slower than the old one, but it works.
+
+Minor changes:
+- [fixed multiple issn issue in SJR](https://github.com/msperlin/GetLattesData/issues/6)
+
+
 ### Version 1.1 (2018-08-19)
 
 Sadly, lattes website is once again using captcha. This version removes the code needed to build the package, so that it can be hosted in CRAN.